This project will be set aside... for now at least
Effect Framework is so giant , complex, tedious. I only managed to
encapsule a few states in this simple Effect Framework. But Effect11
(fx11) is still powerful enough to handle current render engine's
project scale, so it's no need to continue developing this project to
god will know when. But I have a good try to TOKENIZER and PARSER and
that is the biggest gain I'd got during the development of this project.
But I don't want to waste any more time trying to developing such a
giant thing by my own hand ,just because some LINKING ERRORS occur. so
this project will be deprecated until being useful again.